5.1.7 Создаст отдельный компонент CitiesList и вынесет в него отрисовку городов и логику выбора города.
